#410: util: improve file names and split
